Cleanup/Refactor

+++ ioapi_mem.h
@@ -14,19 +14,17 @@
 #ifndef _IOAPI_MEM_H
 #define _IOAPI_MEM_H
 
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include "zlib.h"
 #include "ioapi.h"
+#include "ioapi_mem.h"
+
+#include "zconf.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 voidpf ZCALLBACK fopen_mem_func OF((voidpf opaque,const char* filename,int mode));
-voidpf ZCALLBACK fopendisk_mem_func OF((voidpf opaque, voidpf stream, int number_disk, int mode));
+voidpf ZCALLBACK fopendisk_mem_func OF((voidpf opaque, voidpf stream, unsigned long number_disk, int mode));
 uLong ZCALLBACK fread_mem_func OF((voidpf opaque,voidpf stream,void* buf,uLong size));
 uLong ZCALLBACK fwrite_mem_func OF((voidpf opaque,voidpf stream,const void* buf,uLong size));
 long ZCALLBACK ftell_mem_func OF((voidpf opaque,voidpf stream));
